Refactor into several files for DOM, main logic (logic) and main file (index.js)
